refactor(pipeline/network): extract DiscoveryEngine from NetworkDiscovery
Move the live INetworkDiscovery facade out of NetworkDiscovery.ts
into a dedicated DiscoveryEngine/ sub-module. The engine owns the
captured response pool, the core/endpoint/header/bucketing/lifecycle/
traffic/auth-cache/api-origin/failure-gate method bundles, and the
cross-frame POST interceptor:

  • buildCoreMethods / EndpointMethods / HeaderMethods /
    buildEndpointMethods — the typed bundles that compose
    INetworkDiscovery.
  • interceptPostResponses — page.waitForResponse loop that catches
    POSTs from cross-origin iframes that page.on('response') misses.
  • buildBucketingMethods — pre-nav / post-nav splitter driven by
    the dashboard-click timestamp.
  • INetworkDiscoveryOpts + createNetworkDiscovery — wire everything
    together against a live Playwright Page, honouring the
    isDeferAttach toggle so the homepage / WAF-check window can run
    without attached listeners.

NetworkDiscovery.ts now hosts only createFrozenNetwork +
buildFrozenHeaders (commit 7 handles that final extraction) and
re-exports createNetworkDiscovery from DiscoveryEngine so all 50
existing import sites keep working.

Phase 4 split: commit 6/9. NetworkDiscovery.ts: 568 → 179 LoC.
